refactored database package, minimized code dups between sqlite and postgres implementations
